Added a bug since there were not enough of them to keep the testers busy.

+++ Projects/ETravel.Coffee.Service/Services/OrdersService.cs
@@ -11,6 +11,7 @@ namespace ETravel.Coffee.Service.Services
 {
 	public class OrdersService : RestServiceBase<Orders>
 	{
+		public IOrderItemsRepository OrderItemsRepository { get; set; }
 		public IOrdersRepository OrdersRepository { get; set; }
 
 		public override object OnGet(Orders request)
@@ -57,6 +58,13 @@ namespace ETravel.Coffee.Service.Services
 
 		public override object OnDelete(Orders request)
 		{
+			var orderItems = OrderItemsRepository.ForOrderId(new Guid(request.Id)).ToList();
+			
+			// BUG: If more than one order items are contained in the order, it can 't be deleted.
+			if (orderItems.Count > 1) return new HttpResult { StatusCode = HttpStatusCode.OK };
+
+			orderItems.ForEach(x => OrderItemsRepository.Delete(x.Id.GetValueOrDefault()));
+
 			OrdersRepository.Delete(new Guid(request.Id));
 
 			return new HttpResult { StatusCode = HttpStatusCode.OK };
